refactor(webrtc): extract SDP-negotiation methods from rtc-peer-connection.ts (1039 → 978 + 127)
First slice of the long-deferred `rtc-peer-connection.ts` split. Same
`install*Methods(proto)` pattern as the WebGL2 / canvas2d-core splits
(PRs #273, #262), applied to the `RTCPeerConnection` class.

Extracts the five W3C JSEP entry points that drive offer/answer SDP
negotiation — they form the cleanest extractable group (each is
self-contained, only touches a handful of internal fields, and they
don't call out to other large methods):

  • createOffer            (15 LoC — `create-offer` action signal +
                            `iceRestartNeeded` flag toggle)
  • createAnswer           (8 LoC — `create-answer` action signal)
  • setLocalDescription    (27 LoC — implicit-SLD §4.4.1.6 path +
                            `set-local-description` action signal)
  • setRemoteDescription   (14 LoC — `set-remote-description` action
                            signal + `hasNegotiated` flag set)
  • addIceCandidate        (7 LoC — `add-ice-candidate` action signal)

New module: packages/web/webrtc/src/rtc-peer-connection/sdp-negotiation.ts

Same shape as the previous splits: an `interface SdpNegotiationMethods`
declaration-merged into `RTCPeerConnection` via `declare module`, a
`const sdpNegotiationMethods` object literal carrying the verbatim
method bodies, and an `installSdpNegotiationMethods(proto)` function
that copies them onto the prototype.

To let the extracted methods reach the shared internal state through
their `this: RTCPeerConnection` typing, four formerly-`private` fields
(`_pipeline`, `_webrtcbin`, `_iceRestartNeeded`, `_hasNegotiated`) and
two helpers (`_rejectIfClosed`, `_setStructureField`) are demoted to
package-internal (no `private` modifier; `_`-prefixed naming preserved
as the internal convention). Same approach as WebGL2's `_native2` /
`_vertexArrayObjects` package-internal fields. The remaining state
(`_bridge`, `_conf`, `_closed`, `_dataChannels`, `_transceivers`,
`_senders`, `_receivers`, `_iceTransport`, `_dtlsTransport`,
`_sctpTransport`) stays `private` — none of it is touched by the
extracted methods.

Base file gains the `installSdpNegotiationMethods(RTCPeerConnection.prototype)`
call after the class declaration, plus a deliberate side-effect
`import './rtc-peer-connection/sdp-negotiation.js'` kept separate from
the named import. tsc strips value-only named imports from emitted
`.d.ts`, so the bare side-effect import is required to preserve the
`declare module` augmentations downstream. Unused `RTCIceCandidateInit`
type import dropped from the base.

Public surface preserved verbatim: every export from `@gjsify/webrtc`
works unchanged. Method signatures (`createOffer`, `createAnswer`,
`setLocalDescription`, `setRemoteDescription`, `addIceCandidate`)
identical, including the `_options` ignored parameters (the GIR layer
never reads them — webrtcbin's `create-offer`/`create-answer` action
signals take their options through a `GstStructure`, populated above
from `iceRestartNeeded` only).

Test result: 95 GJS webrtc tests pass / 83 fail — identical to the
pre-refactor baseline (re-run via `git worktree` against `main` to
verify). All 83 failures are pre-existing, caused by
`libnice-gstreamer1` not being installed on this dev box; the
webrtcbin constructor's `ensureWebrtcbinAvailable()` throws
NotSupportedError before any SDP method runs. The SDP extraction is
exercised at compile-time via tsc (clean) and at runtime via the
prototype-install side-effect being preserved in the emitted
`.d.ts`.

Larger remaining method groups left in the base file for follow-up:

  • addTransceiver         (~130 LoC — touches `_transceivers`,
                            `_senders`, `_receivers`, the transport
                            lazy-init, plus webrtcbin's
                            `add-transceiver` / `request_pad_simple`
                            duality)
  • createDataChannel      (~95 LoC — touches `_dataChannels`, the
                            `_ensureSctpTransport` path, and the
                            `create-data-channel` action signal)
  • addTrack / removeTrack (~50 LoC — touches `_senders` reuse logic)
  • getStats / restartIce / setConfiguration / getConfiguration
    (~80 LoC combined — each touches a different mix of internal
    state; `restartIce` also calls the private
    `_handleNegotiationNeeded` signal handler)

Each can use the same pattern in a future PR once it's worth the
prototype-install round-trip; this PR intentionally stops at the
isolated JSEP-method group to keep the demoted-field surface small.

STATUS.md: new Completed entry.
