+++ STATUS.md
@@ -405,6 +405,8 @@ Not yet implemented (but potentially relevant for GJS projects):
 
 ### Completed
 
+- ~~**`@gjsify/webrtc` — extract SDP-negotiation methods from `rtc-peer-connection.ts` (1039 → 978 + 127)**~~✓ — First slice of the deferred `rtc-peer-connection.ts` split. Same `install*Methods(proto)` pattern as the WebGL2 / canvas2d-core splits (PRs #273, #262). The five JSEP entry points (`createOffer`, `createAnswer`, `setLocalDescription`, `setRemoteDescription`, `addIceCandidate`) move to a sibling `packages/web/webrtc/src/rtc-peer-connection/sdp-negotiation.ts` module that declares an `interface SdpNegotiationMethods` augmented onto `RTCPeerConnection` via `declare module`. Base file `rtc-peer-connection.ts` shrinks from 1039 → 978 LoC; method bodies move verbatim. To let the extracted methods reach the shared internal state through their `this: RTCPeerConnection` typing, four formerly-`private` fields (`_pipeline`, `_webrtcbin`, `_iceRestartNeeded`, `_hasNegotiated`) and two helpers (`_rejectIfClosed`, `_setStructureField`) are demoted to package-internal (no `private`, `_`-prefixed) — same convention as the WebGL2 `_native2` / `_vertexArrayObjects` package-internal fields. Base file calls `installSdpNegotiationMethods(RTCPeerConnection.prototype)` after class declaration plus a deliberate side-effect `import './rtc-peer-connection/sdp-negotiation.js'` (kept separate from the named import so tsc preserves it in emitted `.d.ts`). Public surface preserved verbatim: every export from `@gjsify/webrtc` works unchanged, class method signatures identical. 95 GJS webrtc tests pass / 83 fail — identical to pre-refactor baseline (all 83 failures are pre-existing, due to missing `libnice-gstreamer1` on the dev box; the webrtcbin construction path throws before any SDP method runs, so the SDP extraction is exercised at compile-time via tsc + at runtime via the prototype-install side-effect). Other large method groups left in the base file for follow-up: `addTransceiver` (~130 LoC), `createDataChannel` (~95 LoC), the `add/removeTrack` family, and `getStats`/`restartIce`/`setConfiguration` — each can use the same pattern once a future PR decides it's worth the prototype-install round-trip.
+
 - ~~**`as any` cleanup batch — long-tail edition (25 → 0 real casts across 5 files)**~~✓ — Fifth batch in the `as any` reduction sweep, targeting the 5-cast long-tail. Each file gets a typed-view (`interface _Globals` for register modules, `Parameters<EventTarget['addEventListener']>` for listener boundary types, `_ZlibSync` / `_GjsRuntimeGlobals` for cross-runtime imports). Files touched (before → after): `packages/web/message-channel/src/index.ts` 5 → 0 (`Parameters<EventTarget['addEventListener']>[1]` aliases for the `_onmessage` / `_onmessageerror` accessors; the override-method signature stays `any` because lib.dom and `@gjsify/dom-events` have unrelated `EventListenerOrEventListenerObject` types and the public surface must match lib.dom), `packages/web/gamepad/src/register.ts` 5 → 0 (`interface _GamepadGlobals { navigator?: { getGamepads?: ... }; GamepadEvent?: ... }`), `packages/web/compression-streams/src/index.ts` 5 → 0 (`interface _ZlibSync` for the dynamic `node:zlib` import; the `globalThis.CompressionStream` / `DecompressionStream` reads cast through the declared local-impl type), `packages/node/ws/src/websocket.ts` 5 → 0 (`NativeWebSocketCtor` structural type for the `@gjsify/websocket` constructor that accepts the Soup-upgrade options arg; the `Buffer.isBuffer(x as any)` casts dropped because Buffer.isBuffer's `unknown` signature already accepts the union), `packages/node/sab-native/src/ts/index.ts` 5 → 0 (`interface _GjsRuntimeGlobals` capturing `imports.{gi, byteArray}` + `Buffer` for the GJS-runtime bootstrap reads that pre-date `@girs/*` resolution). tsc clean across all 5 packages; tests still green: message-channel 16, gamepad ✓, compression-streams 29, ws 43, sab-native 58. Total: 25 real `as any` casts removed; ~50 LoC added in typed-view declarations.
 
 - ~~**`@gjsify/worker_threads` — split `src/worker.ts` into per-concern modules (873 → 552 + 340)**~~✓ — Same per-concern split pattern. Base `worker.ts` shrinks from 873 LoC to 552 LoC; the parent-side `Worker` class + WorkerOptions interface + IPC plumbing (spawn, postMessage, terminate, stdin/stdout/stderr readers, `_extractCrossProcessPorts`, `_serializeWithSabTransfer`, `_resolveFilename`, `_cleanup`) stay together because they share tight coupling with the per-instance `Gio.Subprocess`, the `_sabSocketFd`, and the `_portRegistry: CrossProcessPortRegistry`. The clear extraction target is the ~330-LoC `BOOTSTRAP_CODE` JS-as-string template — the entire JS body that runs in the spawned gjs child (`import GLib from 'gi://GLib'`, the `MainLoop`, async stdin reader, `materialise()` walker, `makeSharedBuffer()` / `makeChildSidePort()` / `parentPort` factories, the SCM_RIGHTS fd-drain logic) lives in its own runtime; the parent class never reads it, only encodes the string and hands the path to `Gio.SubprocessLauncher.spawnv`. New module `packages/node/worker_threads/src/worker-bootstrap.ts` (340 LoC including its file header + the verbatim preserved comment block explaining the side-channel ordering and the lazy `imports.gi.GjsifySabNative` trick) exports a single `BOOTSTRAP_CODE` const. The base file imports it via `import { BOOTSTRAP_CODE } from './worker-bootstrap.js';` — `BOOTSTRAP_BYTES = _encoder.encode(BOOTSTRAP_CODE)` stays in `worker.ts` next to its sole consumer (the temp-file write in the `Worker` constructor). Public surface preserved verbatim; no exports changed. 286 GJS + 281 Node worker_threads tests still green; tsc + build clean. Same approach as the `@gjsify/buffer` `buffer-codec.ts` extraction (PR #283) — pull out a self-contained subprocess-side concern that has zero state-coupling to the parent class, so the base file becomes easier to reason about without losing any expressive power.
